+++ account/templates/account/signin.html
@@ -15,7 +15,7 @@
                     class="img-fluid" alt="Sample image">
             </div>
             <div class="col-md-8 col-lg-6 col-xl-4 offset-xl-1">
-                <form method="POST" action="{% url 'main_page:signin' %}">
+                <form method="POST" action="{% url 'account:signin' %}">
                     <div class="d-flex flex-row align-items-center justify-content-center justify-content-lg-start">
                         <p class="lead fw-normal mb-0 me-3">Sign in</p>
                     </div>
@@ -48,25 +48,29 @@
                     <div class="text-center text-lg-start mt-4 pt-2">
                         <button type="submit" class="btn btn-primary btn-lg"
                             style="padding-left: 2.5rem; padding-right: 2.5rem;">Login</button>
-                        <p class="small fw-bold mt-2 pt-1 mb-0">Don't have an account? <a href="/main_page/signup"
+                        <p class="small fw-bold mt-2 pt-1 mb-0">Don't have an account? <a href="/account/signup"
                                 class="link-danger">Register</a></p>
                     </div>
 
+                    <script>
+                        window.setTimeout(function () {
+                            $(".alert-auto-dismissible").fadeTo(500, 0).slideUp(500, function () {
+                                $(this).remove();
+                            });
+                        }, 4000);
+                    </script>
                     <div class="col-sm-12">
                         {% if messages %}
                         <div class="messages">
                             {% for message in messages %}
-                            <div class="alert alert-info">
+                            <div class="alert {{ message.tags }} alert-auto-dismissible alert-dismissible notification-container text-center"
+                                role="alert">
                                 {{ message.message }}
                                 {{ message.tags }}
                             </div>
                             {% endfor %}
                         </div>
                         {% endif %}
-
-
-                        <hr />
-                    </div>
                 </form>
             </div>
         </div>
