+++ account/templates/account/signup.html
@@ -15,7 +15,7 @@
                     class="img-fluid" alt="Sample image">
             </div>
             <div class="col-md-8 col-lg-6 col-xl-4 offset-xl-1">
-                <form method="POST" action="{% url 'main_page:signup' %}">
+                <form method="POST" action="{% url 'account:signup' %}">
                     {% csrf_token %}
                     <div class="d-flex flex-row align-items-center justify-content-center justify-content-lg-start">
                         <p class="lead fw-normal mb-0 me-3">Create Account</p>
@@ -33,17 +33,28 @@
                             placeholder="Enter password" name="password" />
                         <label class="form-label" for="form3Example4">Password</label>
                     </div>
-                   
+
                     <div class="text-center text-lg-start mt-4 pt-2">
                         <button type="submit" class="btn btn-primary btn-lg"
                             style="padding-left: 2.5rem; padding-right: 2.5rem;">Create</button>
 
                     </div>
+                    <script>
+                        window.setTimeout(function () {
+                                $(".alert-auto-dismissible").fadeTo(500, 0).slideUp(500, function () {
+                                    $(this).remove();
+                                });
+                            }, 4000);
+                    </script>
                     <div class="col-sm-12">
                         {% if messages %}
                         <div class="messages">
                             {% for message in messages %}
-                            <div class="alert alert-info">
+                            <div class="alert {{ message.tags }} alert-auto-dismissible alert-dismissible notification-container text-center"
+                                role="alert">
+                                <button type="button" class="close" data-dismiss="alert" aria-label="Close">
+                                    <span aria-hidden="true">&times;</span>
+                                </button>
                                 {{ message.message }}
                                 {{ message.tags }}
                             </div>
