+++ main_page/models.py
@@ -1,21 +1,17 @@
 from django.db import models
 import base64
 
-class User(models.Model): 
-    id=models.CharField(max_length=50,primary_key=True);
-    password=models.TextField();
-
 class Lyrics(models.Model):
     id=models.AutoField(primary_key=True);
     lyrics=models.TextField();
-    userid=models.ForeignKey("User",on_delete=models.CASCADE)
+    userid=models.ForeignKey("account.User",on_delete=models.CASCADE)
     def __str__(self):
         return self.lyrics;
     
 class Compose(models.Model):
      id=models.AutoField(primary_key=True);
      music=models.TextField();
-     adminid=models.ForeignKey("User",on_delete=models.CASCADE)
+     adminid=models.ForeignKey("account.User",on_delete=models.CASCADE)
 
 class Images(models.Model):
     id=models.AutoField(primary_key=True);
@@ -28,5 +24,5 @@ class Images(models.Model):
     createdAt=models.DateTimeField();
     updatedAt=models.DateTimeField();
     imagescol=models.CharField(max_length=45);
-    memberid=models.ForeignKey("User",on_delete=models.CASCADE);
+    memberid=models.ForeignKey("account.User",on_delete=models.CASCADE);
