+++ meta4music/urls.py
@@ -20,4 +20,5 @@ from django.urls import path,include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('main_page/',include('main_page.urls')),
+    path('account/',include('account.urls'))
 ]
